Reinstate Guava dependency
This fixes an error caused by conflicting merges. A PR added new code
which depended on Guava while another deleted the dependency.
The commit adds the dependency rather than updating the code to not
require the dependency because my understanding is that Guava immutable
sets provide better guarantees than Java unmodifiable collections.
Specifically, if a reference is kept to the set used to construct an
unmodifiable set, then any mutation to that set is reflected in the
unmodifiable set. The Guava immutable set creates a copy.